fixe bytes array to uint16 * and add MACRO FOREACH_OPCODES

+++ src/rom_parsing/binary_rom_parser.c
@@ -5,7 +5,7 @@
 #include <sys/mman.h>
 #include <sys/stat.h>
 
-uint8_t *parse_binary_rom(int fd, uint32_t *num_bytes) {
+uint16_t *parse_binary_rom(int fd, uint32_t *num_bytes) {
   struct stat st;
 
   if (fstat(fd, &st) == -1) {
@@ -15,7 +15,7 @@ uint8_t *parse_binary_rom(int fd, uint32_t *num_bytes) {
 
   *num_bytes = st.st_size;
 
-  uint8_t *file_addr = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
+  uint16_t *file_addr = mmap(NULL, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0);
   if (file_addr == MAP_FAILED) {
     critical_error("(binary_rom_parser::parse_binary_rom) : Error mmap file");
   }
